MOIntegralHelper create to manage AO -> Mo transformations. CC refactorization started. Single precision code is deprecated and will need an update from Lints.jl
