+++ src/Core/Options.jl
@@ -53,9 +53,10 @@ const Default = Dict{String,Union{Float64,Int,String,Bool}}(
                                   "unit" => "angstrom",
                                   "reference" => "rhf",
                                   "scf_max_iter" => 50,
-                                  "scf_max_rms" => 10^-10,
-                                  "scf_alg" => "conventional",
-                                  "mp2_alg" => "conventional",
+                                  "scf_max_rms" => 10^-9,
+                                  "scf_e_conv" => 10^-10,
+                                  "scf_alg" => 1,
+                                  "mp2_alg" => 1,
                                   "df" => false, 
                                   "oda" => true,
                                   "oda_cutoff" => 1E-1,
@@ -63,14 +64,13 @@ const Default = Dict{String,Union{Float64,Int,String,Bool}}(
                                   "scf_guess" => "gwh",
                                   "quiet" => true,
                                   "precision" => "double",
-                                  "cc_alg" => "CTF",
+                                  "cc_alg" => 1,
                                   "ci_alg" => "aci",
                                   "det_size" => 64,
                                   "σ"      => 0.001,
                                   "γ"      => 1.0,
                                   "ζ"      => 0.95,
                                   "ζsize"  => 0,
-                                  "e_conv" => 10,
                                   "d_conv" => 8,
                                   "cc_max_iter" => 50,
                                   "cc_max_rms" => 10^-10,
