+++ src/Fermi.jl
@@ -19,7 +19,6 @@ include("Core/AuxiliarStructs.jl")
 include("Backend/Integrals.jl")
 include("Methods/HartreeFock/HartreeFock.jl")
 include("Methods/MollerPlesset/MollerPlesset.jl")
-#include("Methods/ConfigurationInteraction/ConfigurationInteraction.jl")
 #include("Methods/CoupledCluster/CoupledCluster.jl")
 #include("Methods/FocalPointAnalysis/FocalPointAnalysis.jl")
