+++ src/Methods/CoupledCluster/RCCSD/RCCSD.jl
@@ -1,5 +1,22 @@
+using Fermi.HartreeFock.RHF
+
+export RCCSD
+
+abstract type RCCSDAlgorithm end
+
+function get_rccsd_alg()
+    implemented = [RCCSD1()]
+    N = Options.get("cc_alg")
+    try 
+        return implemented[N]
+    catch BoundsError
+        throw(InvalidFermiOption("implementation number $N not available for RCCSD."))
+    end
+end
+
 """
     Fermi.CoupledCluster.RCCSD
+    TODO
 
 Fermi struct that holds information about RCCSD wavefunctions
 
@@ -13,47 +30,58 @@ _struct tree:_
 
 **RCCSD** <: AbstractCCWavefunction <: AbstractCorrelatedWavefunction <: AbstractWavefunction
 """
-struct RCCSD{T} <: AbstractCCWavefunction
+struct RCCSD{T} <: AbstractCCWavefunction 
     GuessEnergy::T
     CorrelationEnergy::T
-    T1::_T where _T <: Fermi.AbstractTensor
-    T2::_T where _T <: Fermi.AbstractTensor
+    T1::AbstractArray{T,2}
+    T2::AbstractArray{T,4}
     converged::Bool
 end
 
-# Most general function. It defines the precision and call a precision-specific function
-"""
-    Fermi.CoupledCluster.RCCSD()
-
-Compute a RCCSD wave function using Fermi.CurrentOptions data.
-"""
-function RCCSD()
-    prec = select_precision(Fermi.CurrentOptions["precision"])
-    RCCSD{prec}()
+function RCCSD(x...)
+    precision = Options.get("precision")
+    if precision == "single"
+        RCCSD{Float32}(x...)
+    elseif precision == "double"
+        RCCSD{Float64}(x...)
+    else
+        throw(InvalidFermiOption("precision can only be `single` or `double`. Got $precision"))
+    end
 end
 
-function RCCSD(refwfn::RHF)
-    prec = select_precision(Fermi.CurrentOptions["precision"])
-    alg = select_algorithm(Fermi.CurrentOptions["cc_alg"])
-    dummy = RCCSD{prec}(0.0,0.0,Fermi.MemTensor(zeros(prec,0,0)),Fermi.MemTensor(zeros(prec,0,0,0,0)),false)
-    RCCSD{prec}(refwfn, dummy, alg)
+function RCCSD{Float64}(mol = Molecule(), ints = IntegralHelper{Float64}())
+
+    # Compute Restricted Hartree-Fock
+    refwfn = RHF(mol, ints)
+
+    # Delete integrals that are not gonna be used anymore
+    delete!(ints, "S", "T", "V", "JKERI")
+    RCCSD{Float64}(refwfn, ints)
 end
 
-function RCCSD(refwfn::RHF, guess::RCCSD)
-    prec = select_precision(Fermi.CurrentOptions["precision"])
-    alg = select_algorithm(Fermi.CurrentOptions["cc_alg"])
-    RCCSD{prec}(refwfn, guess, alg)
+function RCCSD{Float32}(mol = Molecule(), ints = IntegralHelper{Float32}())
+
+    # Note that using this method a new IntegralHelper object
+    # is created within the RHF call. This is necessary becasue
+    # RHF needs a double precision integral helper
+
+    # Compute Restricted Hartree-Fock
+    refwfn = RHF(mol)
+
+    RCCSD{Float32}(refwfn, ints)
 end
 
-function RCCSD(guess::RCCSD{T}) where T <: AbstractFloat
-    prec = eltype(guess.T2.data)
-    RCCSD{prec}(guess)
+function RCCSD{T}(refwfn::RHF, ints::IntegralHelper{T})
+
+    # Create zeroed guesses for amplitudes
+
+    o = refwfn.ndocc - Options.get("drop_occ")
+    v = refwfn.nvir - Options.get("drop_nvir")
+    T1guess = FermiMDzeros(T, o, v)
+    T2guess = FermiMDzeros(T, o, o, v, v)
+    RCCSD{T}(refwfn, ints, T1guess, T2guess, alg = get_rccsd_alg())
 end
 
-function RCCSD{T}() where T <: AbstractFloat
-    dummy = RCCSD{T}(0.0,0.0,Fermi.MemTensor(zeros(T,0,0)),Fermi.MemTensor(zeros(T,0,0,0,0)),false)
-    RCCSD{T}(dummy)
-end
 
 """
     Fermi.CoupledCluster.RCCSD{T}()
@@ -73,3 +101,6 @@ include("DF-CTF.jl")
 # main kernel (ccsd iteration logic)
 include("kernel.jl")
 
+# For each implementation a singleton type must be create
+struct RCCSD1 <: RCCSDAlgorithm end
+include("RCCSDa.jl")
